requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
celery==5.3.4
redis==4.5.5
stripe==5.5.0
reportlab==4.0.4
django-import-export==3.3.0
drf-yasg==1.21.5
pyotp==2.8.0
twilio==8.8.0